repo/requirements.txt
# Advanced ML Requirements for Production-Grade Invoice AI
streamlit>=1.28.0
pandas>=1.5.0
plotly>=5.10.0
numpy>=1.21.0
scikit-learn>=1.1.0
Pillow>=9.0.0
PyPDF2>=3.0.0
pytesseract>=0.3.10
opencv-python>=4.7.0
regex>=2022.7.9
pyyaml>=6.0

# Advanced ML Libraries
torch>=1.12.0
transformers>=4.21.0
sentence-transformers>=2.2.0
spacy>=3.4.0
dateparser>=1.1.0
fuzzywuzzy>=0.18.0
python-levenshtein>=0.20.0

# Deep Learning & NLP
tensorflow>=2.10.0
keras>=2.10.0
huggingface-hub>=0.10.0
tokenizers>=0.13.0

# Advanced Text Processing
nltk>=3.7
textblob>=0.17.1
langdetect>=1.0.9
textdistance>=4.5.0

# Computer Vision for OCR
easyocr>=1.6.0
paddleocr>=2.6.0

# ML Utilities
joblib>=1.2.0
xgboost>=1.6.0
lightgbm>=3.3.0
catboost>=1.1.0

# Feature Engineering
feature-engine>=1.5.0
category-encoders>=2.5.0

# Model Monitoring
mlflow>=2.0.0
wandb>=0.13.0
